feat(signature-native): fix imports / review
